+++ turbopack/crates/turbo-tasks-macros-tests/tests/test_config.trs
@@ -1,0 +1,1 @@
+turbo_tasks::TurboTasks::new(turbo_tasks_memory::MemoryBackend::new(usize::MAX))
